+++ src/interface/mod.rs
@@ -1,6 +1,6 @@
 //! Interface trait
-pub mod spi;
-pub use self::spi::SpiInterface;
+//pub mod spi;
+//pub use self::spi::SpiInterface;
 pub mod i2c;
 pub use self::i2c::I2cInterface;