+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.3] - 2022-01-28
+
+### Removed
+- SPI mode temporarily disabled due to multibyte read problems
+
+### Fixed
+- bug in the `interrupt.rs` configuration settings
+
 ## [0.1.2] - 2022-01-24
 
 ### Changed